+++ lizard_auth_server/admin.py
@@ -250,8 +250,7 @@ class PortalAdmin(admin.ModelAdmin):
 class OrganisationAdmin(admin.ModelAdmin):
     model = models.Organisation
     search_fields = ['name']
-    list_display = ['name', 'already_migrated', 'num_user_profiles',
-                    'num_roles']
+    list_display = ['name', 'num_user_profiles', 'num_roles']
     readonly_fields = ['unique_id']
     inlines = [OrganisationRoleInline]